+++ client/src/components/Accommodation/GoogleMapsHotelCard.js
