+++ client/src/components/Trip/TripDisplay.css
